First pass at migrating smoke tests to pytest
