+++ tests/e2e/test_hermes_config_mapping.py
@@ -5,29 +5,23 @@
 
 from __future__ import annotations
 
-import sys
-import tempfile
 from pathlib import Path
 
 import yaml
 
-ROOT = Path(__file__).resolve().parents[1]
-ADAPTERS_COMMON_SRC = ROOT / "adapters" / "common" / "src"
 
-def main() -> None:
-    hermes_common, _common_utils = load_common()
-    with tempfile.TemporaryDirectory(prefix="fabric-hermes-config-") as tmpdir:
-        hermes_home = Path(tmpdir) / "home"
-        config_path, config = hermes_common.write_hermes_config(
-            payload(tmpdir),
-            hermes_home,
-            relay_enabled=True,
-        )
+def test_hermes_config_mapping(hermes_common, tmp_path: Path):
+    hermes_home = tmp_path / "home"
+    config_path, config = hermes_common.write_hermes_config(
+        payload(str(tmp_path)),
+        hermes_home,
+        relay_enabled=True,
+    )
 
-        assert config_path == hermes_home / "config.yaml"
-        assert config_path.is_file()
-        saved = yaml.safe_load(config_path.read_text())
-        assert saved == config
+    assert config_path == hermes_home / "config.yaml"
+    assert config_path.is_file()
+    saved = yaml.safe_load(config_path.read_text())
+    assert saved == config
 
     assert config["model"] == {
         "provider": "nvidia",
@@ -45,16 +39,6 @@ def main() -> None:
     }
     assert config["platform_toolsets"]["cli"] == []
     assert config["plugins"]["enabled"] == ["observability/nemo_relay"]
-
-def load_common():
-    common_src = ADAPTERS_COMMON_SRC.as_posix()
-    if common_src not in sys.path:
-        sys.path.insert(0, common_src)
-
-    import nemo_fabric_adapters.common.hermes as hermes_common
-    import nemo_fabric_adapters.common.utils as common_utils
-
-    return hermes_common, common_utils
 
 
 def payload(tmpdir: str) -> dict:
@@ -89,7 +73,3 @@ def payload(tmpdir: str) -> dict:
             }
         },
     }
-
-
-if __name__ == "__main__":
-    main()
