+++ tests/e2e/test_hermes_session.py
@@ -7,52 +7,48 @@ Drives ``FabricClient.start -> invoke -> invoke -> stop`` against the Hermes SDK
 and CLI adapters and asserts the session carries conversation memory across
 turns through the same Fabric runtime handle.
 
-Unlike ``smoke_hermes_sdk.py`` (which shells out to the CLI), this exercises the
+Unlike ``test_hermes_sdk.py`` (which shells out to the CLI), this exercises the
 SDK session APIs through the native Fabric runtime lifecycle, so this must be
 executed by an interpreter that has BOTH the nemo_fabric native extension and
 Hermes importable:
 
     RUN_FABRIC_HERMES_INTEGRATION=1 NVIDIA_API_KEY=... \\
-        <hermes-venv>/bin/python tests/smoke_hermes_session.py
+        <hermes-venv>/bin/python -m pytest tests/e2e/test_hermes_session.py
 """
 
 from __future__ import annotations
 
-import asyncio
 import importlib.util
 import os
 import sys
 from pathlib import Path
 
-ROOT = Path(__file__).resolve().parents[1]
-sys.path.insert(0, str(ROOT / "python" / "src"))
+import pytest
+
+ROOT = Path(__file__).resolve().parents[2]
 
 
-def main() -> None:
+async def test_hermes_session():
     if os.environ.get("RUN_FABRIC_HERMES_INTEGRATION") != "1":
-        print("skipping: set RUN_FABRIC_HERMES_INTEGRATION=1 to run")
-        return
+        pytest.skip("set RUN_FABRIC_HERMES_INTEGRATION=1 to run")
     if not os.environ.get("NVIDIA_API_KEY"):
-        raise SystemExit("NVIDIA_API_KEY is required")
+        pytest.fail("NVIDIA_API_KEY is required")
     if importlib.util.find_spec("nemo_fabric._native") is None:
-        print(
+        pytest.skip(
             "skipping: the SDK session path needs the nemo_fabric native extension "
             "(pip install -e . into this interpreter)"
         )
-        return
     if importlib.util.find_spec("run_agent") is None:
-        print(
+        pytest.skip(
             "skipping: Hermes (run_agent) is not importable; run with the Hermes "
             "venv python (set HERMES_PYTHON or invoke it directly)"
         )
-        return
     hermes_state_spec = importlib.util.find_spec("hermes_state")
     if hermes_state_spec is None:
-        print(
+        pytest.skip(
             "skipping: Hermes session state (hermes_state) is not importable; run "
             "with the Hermes venv python"
         )
-        return
     hermes_state_origin = hermes_state_spec.origin
     if hermes_state_origin:
         hermes_site_packages = str(Path(hermes_state_origin).resolve().parent)
@@ -63,7 +59,7 @@ def main() -> None:
         )
     python_bin = Path(sys.executable).resolve().parent
     os.environ["PATH"] = f"{python_bin}{os.pathsep}{os.environ.get('PATH', '')}"
-    asyncio.run(_run())
+    await _run()
 
 
 async def _run() -> None:
@@ -81,7 +77,9 @@ async def _run_sdk_session() -> None:
     ) as session:
         assert session.status is SessionStatus.ACTIVE, session.status
 
-        r1 = await session.invoke(input="My name is Robin. Please remember it for later.")
+        r1 = await session.invoke(
+            input="My name is Robin. Please remember it for later."
+        )
         assert r1["status"] == "succeeded", r1
         after_turn1 = session.messages
         assert len(after_turn1) >= 2, after_turn1
@@ -108,7 +106,9 @@ async def _run_cli_session() -> None:
     ) as session:
         assert session.status is SessionStatus.ACTIVE, session.status
 
-        r1 = await session.invoke(input="My name is Robin. Please remember it for later.")
+        r1 = await session.invoke(
+            input="My name is Robin. Please remember it for later."
+        )
         assert r1["status"] == "succeeded", r1
         assert r1["output"]["mode"] == "hermes_cli_session", r1
         assert r1["output"]["session_id"] == session.session_id, r1
@@ -122,8 +122,3 @@ async def _run_cli_session() -> None:
         assert "robin" in response, response
 
     assert session.status is SessionStatus.STOPPED, session.status
-    print("smoke_hermes_session ok")
-
-
-if __name__ == "__main__":
-    main()
