+++ tests/python/test_environment_handle.py
@@ -5,19 +5,15 @@
 
 from __future__ import annotations
 
-import asyncio
 import os
-import sys
 from pathlib import Path
-
-sys.path.insert(0, str(Path(__file__).resolve().parents[1] / "src"))
 
 from nemo_fabric import FabricClient
 
 ROOT = Path(__file__).resolve().parents[2]
 
 
-async def main() -> None:
+async def test_environment_handle():
     async with FabricClient() as client:
         session = await client.start_session(
             ROOT / "examples" / "code-review-agent",
@@ -29,12 +25,7 @@ async def main() -> None:
             await session.stop()
 
     assert os.path.isabs(workspace), f"workspace not absolute: {workspace}"
-    assert (
-        "code-review-agent/examples/code-review-agent" not in workspace
-    ), f"workspace path is doubled: {workspace}"
+    assert "code-review-agent/examples/code-review-agent" not in workspace, (
+        f"workspace path is doubled: {workspace}"
+    )
     assert workspace.endswith("repos/my-service"), workspace
-    print("smoke_environment_handle ok")
-
-
-if __name__ == "__main__":
-    asyncio.run(main())
