+++ tests/python/test_sdk_concurrency.py
@@ -6,33 +6,30 @@
 from __future__ import annotations
 
 import asyncio
-import sys
-import tempfile
 from pathlib import Path
 from shutil import copytree
 
-ROOT = Path(__file__).resolve().parents[2]
-sys.path.insert(0, str(ROOT / "python" / "src"))
-
 from nemo_fabric import FabricClient
 
+ROOT = Path(__file__).resolve().parents[2]
 
-async def run_copy(client: FabricClient, fixture_agent: Path, root: Path, name: str) -> dict:
+
+async def run_copy(
+    client: FabricClient, fixture_agent: Path, root: Path, name: str
+) -> dict:
     agent = root / name
     copytree(fixture_agent, agent)
     return await client.run(agent, profiles=["env_local"], input=f"hello from {name}")
 
 
-async def main() -> None:
+async def test_sdk_concurrency(tmp_path: Path):
     fixture_agent = ROOT / "tests" / "fixtures" / "hermes-shim-agent"
 
     async with FabricClient() as client:
-        with tempfile.TemporaryDirectory(prefix="fabric-sdk-concurrency-") as tmpdir:
-            temp_root = Path(tmpdir)
-            first, second = await asyncio.gather(
-                run_copy(client, fixture_agent, temp_root, "agent-one"),
-                run_copy(client, fixture_agent, temp_root, "agent-two"),
-            )
+        first, second = await asyncio.gather(
+            run_copy(client, fixture_agent, tmp_path, "agent-one"),
+            run_copy(client, fixture_agent, tmp_path, "agent-two"),
+        )
 
     assert first["status"] == "succeeded"
     assert second["status"] == "succeeded"
@@ -41,7 +38,3 @@ async def main() -> None:
     assert first["output"]["received"] == "hello from agent-one"
     assert second["output"]["received"] == "hello from agent-two"
     assert first["artifacts"]["root"] != second["artifacts"]["root"]
-
-
-if __name__ == "__main__":
-    asyncio.run(main())
